add saga for only borrow book with status isReady is true, if isReady is false, it will be roll back

+++ borrowingservice/src/main/java/com/thang/borrowingservice/command/event/BorrowEventHandler.java
@@ -22,6 +22,7 @@ public class BorrowEventHandler {
 
     @EventHandler
     public void on(BorrowDeletedEvent event) {
+        if (borrowRepository.findById(event.getId()).isPresent())
          borrowRepository.deleteById(event.getId());
     }
 }
